% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{AIC.lifelihoodResults}
\alias{AIC.lifelihoodResults}
\title{Akaike information criterion}
\usage{
\method{AIC}{lifelihoodResults}(object, ..., k = length(coef(object)))
}
\arguments{
\item{object}{output of \code{\link[=lifelihood]{lifelihood()}}}

\item{...}{Ignored}

\item{k}{Number of estimated parameter of the modèle. Default to \code{length(coef(object))}}
}
\value{
The AIC
}
\description{
S3 method to compute AIC (Akaike information criterion).
}
\seealso{
\code{\link{BIC}}
}
