% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{AIC}
\alias{AIC}
\alias{AIC.lifelihoodResults}
\title{AIC}
\usage{
\method{AIC}{lifelihoodResults}(object, parameter_name)
}
\arguments{
\item{object}{\code{lifelihoodResults} object from \code{\link[=lifelihood]{lifelihood()}}}

\item{parameter_name}{Name of the parameter. All parameters can be found \href{/articles/2-setting-up-the-configuration-file.html#parameters}{here}.}
}
\value{
The AIC.
}
\description{
S3 method to compute AIC (Akaike information criterion).
}
\seealso{
\code{\link{BIC}}
}
