% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{BIC}
\alias{BIC}
\alias{BIC.lifelihoodResults}
\title{Bayesian information criterion}
\usage{
\method{BIC}{lifelihoodResults}(object)
}
\arguments{
\item{object}{\code{lifelihoodResults} object from \code{\link[=lifelihood]{lifelihood()}}}
}
\value{
The BIC.
}
\description{
S3 method to compute BIC (Akaike information criterion).
}
\seealso{
\code{\link{AIC}}
}
