% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{check_groupby_arg}
\alias{check_groupby_arg}
\title{Check that the \code{groupby} argument is valid for plots}
\usage{
check_groupby_arg(lifelihoodData, groupby)
}
\arguments{
\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}.}

\item{groupby}{A character with one of the covariates passed to \code{\link[=lifelihoodData]{lifelihoodData()}}}
}
\description{
Check that the \code{groupby} argument is valid for plots
}
\keyword{internal}
