% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_lifelihoodData}
\alias{check_lifelihoodData}
\title{Check that an object is of class \code{lifelihoodData}}
\usage{
check_lifelihoodData(object)
}
\arguments{
\item{object}{An object to test.}
}
\description{
Internally, \code{lifelihood} has to check multiple times
that the passed object is the expected one.

It basically raises an explicit error if the object
is not of class \code{lifelihoodData}.
}
\examples{
\dontrun{
# raise an error
obj <- c(1,2,3)
check_lifelihoodData(obj)

# works (does nothing)
class(obj) = "lifelihoodData"
check_lifelihoodData(obj)
}

}
\keyword{internal}
