% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{check_mcmc_consistency}
\alias{check_mcmc_consistency}
\title{Check MCMC consistency}
\usage{
check_mcmc_consistency(object, threshold = 0.05)
}
\arguments{
\item{object}{lifelihoodResults object}

\item{threshold}{Maximum relative difference (default: 5\%)}
}
\value{
TRUE if consistent, FALSE if not
}
\description{
Check if MCMC estimates are consistent with MLE
}
\keyword{internal}
