% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mortality_rate.R
\name{compute_fitted_mortality_rate}
\alias{compute_fitted_mortality_rate}
\title{Compute fitted mortality rate}
\usage{
compute_fitted_mortality_rate(
  lifelihoodResults,
  interval_width,
  event = c("mortality", "maturity"),
  newdata = NULL,
  max_time = NULL,
  groupby = NULL
)
}
\arguments{
\item{lifelihoodResults}{A \code{lifelihoodResults} object}

\item{interval_width}{The interval width used to calculate the
mortality rate. For instance, if the time unit for deaths in
the original dataset is days and \code{interval_width} is set to 10,
the mortality rate will be calculated every 10 days for each group.}

\item{newdata}{Data for computation. If absent, predictions are for
the subjects used in the original fit.}

\item{max_time}{The maximum time for calculating the mortality
rate. If set to NULL, the time of the last observed death is used.}

\item{groupby}{vector of covariate(s) over which mortality rate should be
computed (default is \code{NULL}).
\itemize{
\item If NULL, calculates a single overall mortality rate.
\item If \code{"all"}, calculates mortality rate over each combination
of covariates listed in the\code{lifelihoodData} object provided.
\item Otherwise must be a character (\code{"covariate1"}) or a
character vector (\code{c("covariate1", "covariate2")}).
Note that the function will consider continuous covariates as factors
}}
}
\value{
A dataframe with 3 columns: Interval (time interval, based
on \code{interval_width} value), Group (identifier of a given subgroup,
or "Overall" if groupby = NULL), and MortalityRate (mortality rate
at this time).
}
\description{
Calculate the empirical mortality rate over a given interval
on some new data.
}
