% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compute_mortality_rate}
\alias{compute_mortality_rate}
\title{good function}
\usage{
compute_mortality_rate(
  data,
  start_col,
  end_col,
  covariates,
  interval_width = 2,
  max_time = NULL
)
}
\value{
none
}
\description{
do something
}
