% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_event_rate.R
\name{compute_reproduction_intervals}
\alias{compute_reproduction_intervals}
\title{Compute time interval between clutches}
\usage{
compute_reproduction_intervals(lifelihoodData, verbose = TRUE)
}
\arguments{
\item{lifelihoodData}{Ouput of \code{\link[=lifelihoodData]{lifelihoodData()}}}

\item{verbose}{Boolean whether to print messages during the computation process (default=TRUE)}
}
\value{
A dataframe with time interval between consecutive clutches starting from maturity.
}
\description{
This function computes the time interval between consecutive clutches
and add censored unobserved last clutch between the last clutch and time
of death unless the time of death is right censored. It also removes
individuals that never reproduced.
}
