% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_parameters}
\alias{count_parameters}
\title{Count total number of parameter to fit}
\usage{
count_parameters(x)
}
\arguments{
\item{x}{Config file parsed by \code{\link[yaml:yaml.load]{yaml::yaml.load_file()}}.}
}
\value{
Total number of parameter to fit.
}
\description{
Simple utility function that parses the config yaml file
and count the total number of parameter that will be fit.
}
\keyword{internal}
