% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_lin_decrease_hazard}
\alias{data_lin_decrease_hazard}
\title{Simulated data with a linear decrease in hazard}
\format{
An object of class \code{data.frame} with 100 rows and 20 columns.
}
\usage{
data_lin_decrease_hazard
}
\description{
A simulated life-history data set with sex, treatment, maturity,
reproduction, and death intervals generated under a linear decrease in
hazard model.
}
\keyword{data}
