% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datadaphnia}
\alias{datadaphnia}
\title{Daphnia data}
\format{
An object of class \code{data.frame} with 550 rows and 95 columns.
}
\usage{
datadaphnia
}
\description{
Life-history observations for Daphnia individuals, including genotype,
spore treatment, sex, maturity, reproduction, and death intervals.
}
\keyword{data}
