% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datalenski}
\alias{datalenski}
\title{Lenski data}
\format{
An object of class \code{data.frame} with 18 rows and 27 columns.
}
\usage{
datalenski
}
\description{
Life-history observations with group, sex, maturity, reproduction, and death
intervals.
}
\keyword{data}
