% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datapierrick}
\alias{datapierrick}
\title{Pierrick data}
\format{
An object of class \code{data.frame} with 550 rows and 95 columns.
}
\usage{
datapierrick
}
\description{
Pierrick data
}
\keyword{data}
