% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_param_range}
\alias{delete_param_range}
\title{(internal function) Delete parameter ranges/boundaries file}
\usage{
delete_param_range(file_path)
}
\arguments{
\item{file_path}{Path of where the .txt file is}
}
\description{
(internal function) \code{delete_param_range()} takes a path to the parameter ranges/boundaries file and delete it
}
