% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{delink}
\alias{delink}
\title{Delink function}
\usage{
delink(obs, min, max)
}
\arguments{
\item{obs}{The observed value of the parameter}

\item{min}{The minimum value of the parameter}

\item{max}{The maximum value of the parameter}
}
\value{
Numeric. The transformed parameter
}
\description{
Delink function to transform the parameters from original scale
to lifelihood scale.
}
\keyword{internal}
