% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_os}
\alias{detect_os}
\title{(internal function) find the operating system of the user}
\usage{
detect_os()
}
\value{
string of operatin system
}
\description{
(internal function) \code{detect_os()} finds the OS name
}
\details{
Group of internal functions used in the lifelihood programme
}
