% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_bin.R
\name{execute_bin}
\alias{execute_bin}
\title{Execution of the compiled files}
\usage{
execute_bin(
  path_to_Lifelihood,
  path_input_data,
  path_param_bounds,
  group_by_group,
  MCMC,
  interval,
  SEcal,
  saveprobevent,
  fitness,
  r,
  seed1,
  seed2,
  seed3,
  seed4,
  ntr,
  nst,
  To,
  Tf,
  climbrate,
  precision
)
}
\arguments{
\item{path_to_Lifelihood}{A character string specifying the file path to the compile Lifelihood program (default is NULL)}

\item{path_input_data}{Path to the input text file with the model and data to be fitted.}

\item{path_param_bounds}{Path to the parameter boundaries text file with the min and max boudaries for each parameter.}

\item{group_by_group}{Boolean option to fit the full factorail model with all the interactions between each of the factors}

\item{MCMC}{Perform MCMC sampling of the parameter after convergence to estimate their 95\% confidence interval}

\item{interval}{TBD - Check the actual meaning}

\item{SEcal}{If TRUE, Lifelihood computes the standard error of each parameter using the Hessian matrix (output with value of -1 if standard error cannot be computed due to singularity of the Hessian matrix)}

\item{saveprobevent}{TBD - Check the actual meaning}

\item{fitness}{Reparametrize the model with one parameter as the lifetime reproductive success}

\item{r}{Reparametrize the model with one parameter as the intrinsic rate of increase}

\item{seed1}{First seed number used to reproduce results (same seed = same results).}

\item{seed2}{Second seed number used to reproduce results (same seed = same results).}

\item{seed3}{Third seed number used to reproduce results (same seed = same results).}

\item{seed4}{4th seed number used to reproduce results (same seed = same results).}

\item{ntr}{Number of thread for the paralelisation ?}

\item{nst}{TBD - Check the actual meaning}

\item{To}{Initial temperature for the simulated annealing}

\item{Tf}{Initial temperature for the simulated annealing}

\item{climbrate}{Rate for the simulated annealing ?}

\item{precision}{TBD - Check the actual meaning}
}
\description{
Run lifelihood program in console mode.
}
\keyword{internal}
