% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_to_num}
\alias{factor_to_num}
\title{Convert a vector to factor codes}
\usage{
factor_to_num(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Numeric vector of 0-based factor codes
}
\description{
Convert values to 0-based factor codes for the Pascal program.
}
\keyword{internal}
