% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_effect_names.R
\name{find_event_type}
\alias{find_event_type}
\title{find event type}
\usage{
find_event_type(metric_name)
}
\arguments{
\item{metric_name}{Name of the metric}
}
\value{
Event type: either "reproduction", "maturity" or "mortality"
}
\description{
find event type
}
\keyword{internal}
