% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_effect_names.R
\name{find_parameter_kind}
\alias{find_parameter_kind}
\title{Get Mapping for Element}
\usage{
find_parameter_kind(name)
}
\arguments{
\item{element}{A character string representing the element to be mapped.}
}
\value{
A named list containing the mapped value for the input element.
}
\description{
This function takes an element as input and returns its corresponding mapping from a predefined list. If the element is not found in the mapping, it throws an error.
}
\details{
Get Mapping for Element
}
\keyword{internal}
