% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_file.R
\name{format_config}
\alias{format_config}
\title{Read and parse the configuration file (YAML).}
\usage{
format_config(config, covariates, covar_types)
}
\arguments{
\item{config}{A complete configuration list.}

\item{covariates}{Vector containing the names of the covariates.}

\item{covar_types}{Vector containing the types of the covariates
(either "cat" for categorical or "num" for numerical).}
}
\value{
A character vector that will be used under the model tag in the input text file.
}
\description{
Safely access the configuration file to use for lifelihood.
This function is used in \code{\link[=lifelihood]{lifelihood()}} when creating the
input text file.
}
\keyword{internal}
