% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{format_dataframe_to_txt}
\alias{format_dataframe_to_txt}
\title{Create the input data file from a dataframe}
\usage{
format_dataframe_to_txt(
  df,
  sex,
  sex_start,
  sex_end,
  maturity_start,
  maturity_end,
  matclutch,
  matclutch_size,
  clutchs,
  death_start,
  death_end,
  covariates,
  dist,
  config,
  temp_dir
)
}
\arguments{
\item{df}{Dataframe with the data of life history. It should
have one row per life history / observation.}

\item{sex}{Column name containing the sex of the observations.}

\item{sex_start}{Column name containing the first date of the
interval in which the sex was determined.}

\item{sex_end}{Column name containing the second date of the
interval in which the sex was determined.}

\item{maturity_start}{Column name containing the first date of
the interval in which the maturity was determined.}

\item{maturity_end}{Column name containing the second date of
the interval in which the maturity was determined.}

\item{matclutch}{Whether the maturity event (designated by
\code{maturity_start} and \code{maturity_end}) is a clutch event or not.}

\item{matclutch_size}{Column name containing the size of the
clutch for the maturity event. Only used (and required) if
\code{matclutch} is \code{TRUE}.}

\item{clutchs}{Vector containing the names of the clutch columns.}

\item{death_start}{Column name containing the first date of the
interval in which the death was determined.}

\item{death_end}{Column name containing the second date of the
interval in which the death was determined.}

\item{covariates}{Vector containing the names of the covariates.}

\item{dist}{One-row data frame with columns \code{mortality}, \code{maturity}, and
\code{reproduction}, specifying the model being fitted.}

\item{config}{A complete configuration list.}

\item{temp_dir}{Name of the temporary directory with temporary
files.}
}
\description{
Takes a dataframe (\code{df} argument in \code{\link[=as_lifelihoodData]{as_lifelihoodData()}} function)
and apply to each row the \code{\link[=format_row]{format_row()}} function to create the
input data file.
}
\keyword{internal}
