% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{format_row}
\alias{format_row}
\title{format a single row for the input data file}
\usage{
format_row(row, extra1, extra2, extra3)
}
\arguments{
\item{row}{a row of a dataframe object}

\item{extra1}{name of the first column to add in the input data file}

\item{extra2}{name of the second column to add in the input data file}

\item{extra3}{name of the third column to add in the input data file}
}
\value{
a string of the well formated row
}
\description{
takes a row from a dataframe with input data (sex, maturity, clutch date and death) in interval format and transform it into a large string (required format for the data input file)
}
\details{
Functions used to create input data (.txt) file for the lifelihood program
}
