% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_clutch_vector}
\alias{generate_clutch_vector}
\title{Utility to generate vector of clutch names.}
\usage{
generate_clutch_vector(N)
}
\arguments{
\item{N}{Number of clutch slots to include in the returned vector.}
}
\description{
Utility to generate vector of clutch names.
}
