% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_effects}
\alias{get_effects}
\title{(internal function) Get the estimation in the output file}
\usage{
get_effects(lines, group_by_group = FALSE)
}
\arguments{
\item{lines}{Vector of the output file, where each element is a line of the file}

\item{group_by_group}{Boolean indicating whether the analysis should be made group by group or not (default to false)}
}
\value{
A dataframe of the parsed effects estimation
}
\description{
(internal function) Find the estimated effects in the output file of the lifelihood program
}
