% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_likelihood}
\alias{get_likelihood}
\title{(internal function) get the likelihood in the output file}
\usage{
get_likelihood(lines, group_by_group = FALSE)
}
\arguments{
\item{lines}{vector of the output file, where each element is a line of the file}

\item{group_by_group}{boolean indicating whether the analysis should be made group by group or not (default to false)}
}
\value{
the parsed likelihood
}
\description{
(internal function) find the likelihood value (optimum found) in the output file of the lifelihood program
}
