% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_ratio_max}
\alias{get_ratio_max}
\title{Get the ratio max in the output file}
\usage{
get_ratio_max(lines)
}
\arguments{
\item{lines}{vector of the output file, where each element is a line of the file}
}
\description{
find the ratio max value in the output file of the lifelihood program
}
