% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{has_valid_factor_levels}
\alias{has_valid_factor_levels}
\title{Check for valid factor levels}
\usage{
has_valid_factor_levels(df_train, newdata, covariates)
}
\arguments{
\item{df_train}{Training set passed to \code{\link[=lifelihoodData]{lifelihoodData()}} (\code{df} arg).}

\item{newdata}{New data passed to \code{\link[=predict]{predict()}} (\code{newdata} arg).}

\item{covariates}{Covariates passed to \code{\link[=lifelihoodData]{lifelihoodData()}} (\code{covariates} arg).}
}
\value{
TRUE if all factor levels of each covariate in \code{newdata} passed to \code{\link[=predict]{predict()}} are present in the training data, FALSE otherwise.
}
\description{
The purpose of this function is to ensure that when a user makes a prediction with \code{\link[=predict]{predict()}}, the \code{newdata} contains the same factor levels for its covariates as the training data.

If any levels are missing or mismatched, it raises an error and displays a warning.
}
\keyword{internal}
