% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IntX1toX2MortWei}
\alias{IntX1toX2MortWei}
\title{Probability of dying between t and t+dt afer living until t.}
\usage{
IntX1toX2MortWei(t, dt, param1, param2, law)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{ExpLong}{Numeric. The expected longevity}

\item{Shape}{Numeric. The shape parameter}
}
\value{
The survival probability (numeric)
}
\description{
Probability of dying between t and t+dt afer living until t.
}
\keyword{internal}
