% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood_optim.R
\name{lifelihood_optim}
\alias{lifelihood_optim}
\title{Fit multiple times lifelihood}
\usage{
lifelihood_optim(..., n_fit = 3)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=lifelihood]{lifelihood()}}}

\item{n_fit}{Number of times to fit. Default to 3.}
}
\value{
\code{lifelihoodResults}
}
\description{
This function fits \code{\link[=lifelihood]{lifelihood()}} \code{n_fit} times and return
the fit with the best likelihood.
}
