% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Link function}
\usage{
link(estimate, min, max)
}
\arguments{
\item{estimate}{The estimate of the parameter}

\item{min}{The minimum value of the parameter}

\item{max}{The maximum value of the parameter}
}
\value{
The transformed parameter
}
\description{
Link function to transform the parameters from the
lifelihood scale to the original scale
}
\keyword{internal}
