% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{logLik.lifelihoodResults}
\alias{logLik.lifelihoodResults}
\title{Likelihood}
\usage{
\method{logLik}{lifelihoodResults}(object, ...)
}
\arguments{
\item{object}{output of \code{\link[=lifelihood]{lifelihood()}}}

\item{...}{Ignored}
}
\value{
A number with the value of maximum likelihood found.
}
\description{
S3 method to retrieve likelihood from the output of \code{\link[=lifelihood]{lifelihood()}}
}
