% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{Make the design matrix}
\usage{
make_design_matrix(covariates, data)
}
\description{
Create the design matrix for the expected death and
the survival shape.
}
\keyword{internal}
