% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_effect_names.R
\name{map_parameter_name}
\alias{map_parameter_name}
\title{Associate the parameter name with the associated parameter}
\usage{
map_parameter_name(name)
}
\arguments{
\item{name}{A character string representing the element to be mapped.}
}
\value{
A named list containing the mapped value for the input element.
}
\description{
This function takes a name as input and returns its match from a predefined list. If the element is not found in the match, an error is generated. This function is used to determine which parameter a parameter estimate is based on.
}
\examples{
map_parameter_name("eff_expt_death_geno1")
map_parameter_name("eff_ratio_expt_death_geno1")
map_parameter_name("int_sex_ratio")
}
\keyword{internal}
