% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{mcmc.lifelihoodResults}
\alias{mcmc.lifelihoodResults}
\title{MCMC}
\usage{
mcmc.lifelihoodResults(object, ...)
}
\arguments{
\item{object}{output of \code{\link[=lifelihood]{lifelihood()}}}

\item{...}{Ignored}
}
\value{
A covariance matrix
}
\description{
S3 method to retrieve the mcmc
from the output of \code{\link[=lifelihood]{lifelihood()}}
}
