% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{mcmc}
\alias{mcmc}
\title{MCMC}
\usage{
mcmc(object)
}
\arguments{
\item{object}{output of \code{\link[=lifelihood]{lifelihood()}}}
}
\value{
A covariance matrix
}
\description{
Retrieve the mcmc from the output of \code{\link[=lifelihood]{lifelihood()}}
}
