% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\name{mortality_rate}
\alias{mortality_rate}
\title{Compute empirical mortality rate}
\usage{
mortality_rate(lifelihoodData, interval_width, max_time = NULL)
}
\arguments{
\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}.}

\item{interval_width}{The interval width used to calculate the mortality rate. For instance, if the time unit for deaths in the original dataset is days and \code{interval_width} is set to 10, the mortality rate will be calculated every 10 days for each group.}

\item{max_time}{The maximum time for calculating the mortality rate. If set to NULL, the time of the last observed death is used.}
}
\value{
A dataframe with 3 columns: Interval (time interval, based on \code{interval_width} value), Group (identifier of a given subgroup) and MortalityRate (mortality rate of this sub group at this time).
}
\description{
Calculate the empirical mortality rate over a given interval.
}
\examples{
df <- read.csv(here::here("data/fake_sample.csv"))
df$type <- as.factor(df$type)
df$geno <- as.factor(df$geno)

clutchs <- c(
  "clutch_start1", "clutch_end1", "clutch_size1",
  "clutch_start2", "clutch_end2", "clutch_size2"
)

dataLFH <- lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "mor_start",
  death_end = "mor_end",
  covariates = c("geno", "type"),
  model_specs = c("gam", "lgn", "wei")
)

mort_df <- mortality_rate(dataLFH, interval_width = 5)
head(mort_df)
}
