% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{parallel.simulate}
\alias{parallel.simulate}
\title{Parallelized simulations}
\usage{
parallel.simulate(..., nsim, parallel_seed = FALSEj)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=simulate_life_history]{simulate_life_history()}}.}

\item{nsim}{Number of simulations}

\item{parallel_seed}{Logical or integer, seed used for}
}
\description{
Parallelized simulations
}
