% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_elements}
\alias{parse_elements}
\alias{get_seeds}
\alias{get_likelihood}
\alias{get_param_ranges}
\alias{get_effects}
\alias{get_ratio_max}
\alias{parse}
\title{Parse Specific Elements from Text Lines}
\usage{
get_seeds(lines, group_by_group = FALSE)

get_likelihood(lines, group_by_group = FALSE)

get_param_ranges(lines)

get_effects(lines, group_by_group = FALSE)

get_ratio_max(lines)

parse(lines, element, group_by_group = FALSE)
}
\arguments{
\item{lines}{A character vector containing lines of text from which elements are to be parsed.}

\item{group_by_group}{A logical value indicating whether the elements should be grouped. Default is \code{FALSE}.}

\item{element}{A character string specifying the element to be parsed. Possible values are "seeds", "likelihood", "parameter_ranges", "effects", and "ratio_max".}
}
\value{
The parsed elements, which can be numeric vectors, matrices, or data frames, depending on the type of element requested.
}
\description{
This set of functions is designed to parse various elements such as seeds, likelihood, parameter ranges, effects, and ratio max values from a vector of text lines. The main function \code{parse()} is the primary interface for extracting these elements based on the specified type.
}
\examples{
lines <- c("seed1=123 seed2=456 seed3=789 seed4=101112", 
           "group 1 Likelihood_max=0.95", 
           "Parameter_Range_Table", 
           "param1 0.1 1.0", 
           "param2 0.2 2.0", 
           "ratiomax 12345", 
           "eff_param1 0.5 0.05", 
           "eff_param2 1.5 0.15")
parse(lines, "seeds")
parse(lines, "likelihood")
parse(lines, "parameter_ranges")
parse(lines, "effects")
parse(lines, "ratio_max")

}
