% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_emp_mortality_rate}
\alias{plot_emp_mortality_rate}
\title{Display evolution of empirical mortality rate}
\usage{
plot_emp_mortality_rate(
  lifelihoodData,
  interval_width,
  min_time = 0,
  max_time = NULL,
  log_x = FALSE,
  log_y = FALSE
)
}
\arguments{
\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}.}

\item{interval_width}{The interval width used to calculate the mortality rate. For instance, if the time unit for deaths in the original dataset is days and \code{interval_width} is set to 10, the mortality rate will be calculated every 10 days for each group.}

\item{max_time}{The maximum time for calculating the mortality rate. If set to NULL, the time of the last observed death is used.}

\item{log_x}{Determine whether the x-axis should be displayed on a logarithmic scale.}

\item{log_y}{Determine whether the y-axis should be displayed on a logarithmic scale.}
}
\value{
none
}
\description{
Useful function for creating a good-quality line graph of changes in the empirical mortality rate.
}
\details{
This function requires \href{https://ggplot2.tidyverse.org/}{ggplot2} to be installed.
}
\examples{
df <- read.csv(here::here("data/fake_sample.csv"))
df$type <- as.factor(df$type)
df$geno <- as.factor(df$geno)

clutchs <- c(
  "clutch_start1", "clutch_end1", "clutch_size1",
  "clutch_start2", "clutch_end2", "clutch_size2"
)

dataLFH <- lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "mor_start",
  death_end = "mor_end",
  covariates = c("geno", "type"),
  model_specs = c("gam", "lgn", "wei")
)

plot_emp_mortality_rate(dataLFH, interval_width = 2)
}
