% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mortality_rate_emp}
\alias{plot_mortality_rate_emp}
\title{Display evolution of mortality rate}
\usage{
plot_mortality_rate_emp(
  data,
  start_col,
  end_col,
  covariates,
  interval_width = 2,
  max_time = NULL
)
}
\value{
none
}
\description{
Display evolution of mortality rate
}
