% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\name{pred_mortality_rate}
\alias{pred_mortality_rate}
\title{Compute and visualize predicted mortality rate}
\usage{
pred_mortality_rate(
  results_lifelihood,
  covariates,
  data,
  intervals = seq(0, 20, 5)
)
}
\description{
Compute and visualize predicted mortality rate
}
