% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{print_coef_table}
\alias{print_coef_table}
\title{Print simplified coefficient table}
\usage{
print_coef_table(df, digits = 3)
}
\arguments{
\item{df}{Dataframe with effects}

\item{digits}{Number of digits to display}
}
\description{
Helper function to print a clean coefficient table
}
\keyword{internal}
