% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_event_interval_dt}
\alias{prob_event_interval_dt}
\title{Probability of dying between t and t+dt afer living until t.}
\usage{
prob_event_interval_dt(t, dt, param1, param2, family)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{expt_time_to_event}{Numeric. The expected longevity or time to maturity.}

\item{Shape}{Numeric. The shape parameter}
}
\value{
The survival probability (numeric)
}
\description{
Probability of dying between t and t+dt afer living until t.
}
\keyword{internal}
