% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_file.R
\name{R_to_lifelihood}
\alias{R_to_lifelihood}
\title{Convert model description to lifelihod format}
\usage{
R_to_lifelihood(R_format, covariates)
}
\arguments{
\item{R_format}{String representing the covariates to be adjusted. For example, "geno + type" will use the covariates geno and type.}

\item{covariates}{Vector containing the names of covariates.}
}
\description{
(internal function) Function which transforms a character string describing the covariates to be included into a format which the compiled program can understand. For example, \code{"geno + type"} will become \verb{1 2} if \code{"geno"} is the first element of \code{covariables} and \code{"type"} is the second. This function is used to create the model part of the input file.
}
