% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_lifelihood_tempfiles}
\alias{remove_lifelihood_tempfiles}
\title{Remove all lifelihood temporary files}
\usage{
remove_lifelihood_tempfiles(path = ".")
}
\arguments{
\item{path}{Where to look for. Default to current directory.}
}
\description{
By default, \code{\link[=lifelihood]{lifelihood()}} deletes all the temp files
it creates, but users can set \code{delete_temp_files=FALSE}
to keep them.

After multiple runs, there can be lots of temp files.
This function will just remove them.
}
\examples{
remove_lifelihood_tempfiles()

}
