% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lifelihood.R
\name{run_lifelihood}
\alias{run_lifelihood}
\title{Run lifelihood program}
\arguments{
\item{input_file}{file with the model and data to be fitted}

\item{custom_file}{file with the min and max boudaries for each parameter}

\item{GbyG}{Option to fit the full factorail model with all the interactions between each of the factors}

\item{MCMC}{Perform MCMC sampling of the parameter after convergence to estimate their 95\% confidence interval}

\item{interval}{TBD - Check the actual meaning}

\item{SEcal}{Compute the standard error of eahc parameter using the Hessian matrix}

\item{saveprobevent}{TBD - Check the actual meaning}

\item{fitness}{Reparametrize the model with one parameter as the lifetime reproductive success}

\item{r}{Reparametrize the model with one parameter as the intrinsic rate of increase}

\item{seed1}{first seed used so that random values can found again (https://en.wikipedia.org/wiki/Multiply-with-carry_pseudorandom_number_generator)}

\item{seed2}{second seed used so that random values can found again}

\item{seed3}{third seed used so that random values can found again}

\item{seed4}{fourth seed used so that random values can found again}

\item{ntr}{number of thread for the paralelisation ?}

\item{nst}{TBD - Check the actual meaning}

\item{To}{Initial temperature for the simulated annealing}

\item{Tf}{Initial temperature for the simulated annealing}

\item{climbrate}{Rate for the simulated annealing ?}

\item{precision}{TBD - Check the actual meaning}
}
\value{
dataset with the simulated hatch rate
}
\description{
Run lifelihood program in console mode
}
\examples{
library(here)

# path to inputs
input_file = file.path(
   'data', 'raw_data', 'DataPierrick_GroupbyGroup',
   '100\%mort_Pierrick211genoparinteraction.txt'
)
custom_file = file.path('data', 'custom.txt')
# run the program
run_lifelihood(
   input_file = input_file,
   custom_file = custom_file
)
}
