% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lifelihood_console.R
\name{run_lifelihood_console}
\alias{run_lifelihood_console}
\title{Run lifelihood program}
\usage{
run_lifelihood_console(
  infile = "100\%mort_Pierrick211genoparinteraction.txt",
  customfile = "custom.txt",
  GbyG = 0,
  MCMC = 0,
  interval = 25,
  SEcal = 0,
  saveprobevent = 0,
  fitness = 0,
  r = 0,
  seed1 = 12,
  seed2 = 13,
  seed3 = 14,
  seed4 = 15,
  ntr = 2,
  nst = 2,
  To = 50,
  Tf = 1,
  climbrate = 1,
  precision = 0.001
)
}
\arguments{
\item{infile}{file with the model and data to be fitted}

\item{customfile}{file with the min and max boudaries for each parameter}

\item{GbyG}{Option to fit the full factorail model with all the interactions between each of the factors}

\item{MCMC}{Perform MCMC sampling of the parameter after convergence to estimate their 95\% confidence interval}

\item{interval}{TBD - Check the actual meaning}

\item{SEcal}{Compute the standard error of eahc parameter using the Hessian matrix}

\item{saveprobevent}{TBD - Check the actual meaning}

\item{fitness}{Reparametrize the model with one parameter as the lifetime reproductive success}

\item{r}{Reparametrize the model with one parameter as the intrinsic rate of increase}

\item{seed1}{first seed used so that random values can found again (https://en.wikipedia.org/wiki/Multiply-with-carry_pseudorandom_number_generator)}

\item{seed2}{second seed used so that random values can found again}

\item{seed3}{third seed used so that random values can found again}

\item{seed4}{fourth seed used so that random values can found again}

\item{ntr}{number of thread for the paralelisation ?}

\item{nst}{TBD - Check the actual meaning}

\item{To}{Initial temperature for the simulated annealing}

\item{Tf}{Initial temperature for the simulated annealing}

\item{climbrate}{Rate for the simulated annealing ?}

\item{precision}{TBD - Check the actual meaning}
}
\value{
dataset with the simulated hatch rate
}
\description{
Run lifelihood program in console mode
}
\examples{
run_lifelihood_console(infile="100\%mort_Pierrick211genoparinteraction.txt", ntr=10, To=50, Tf=1, seed2=34)
}
