% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_file.R
\name{safe_access}
\alias{safe_access}
\title{safe access}
\usage{
safe_access(config, path)
}
\arguments{
\item{config}{configuration object}

\item{path}{event and parameter to read}
}
\value{
the read value
}
\description{
Safely access elements in config file. This function exists because yaml.load_file() returns NULL when a value is not found instead of raising an error.
}
\keyword{internal}
