% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{summary.lifelihoodResults}
\alias{summary.lifelihoodResults}
\title{Simple summary for lifelihoodResults objects}
\usage{
\method{summary}{lifelihoodResults}(object, digits = 3, ...)
}
\arguments{
\item{object}{Output of \code{\link[=lifelihood]{lifelihood()}}}

\item{digits}{Number of significant digits to display (default: 3)}

\item{...}{Ignored}
}
\value{
Invisibly returns the object
}
\description{
S3 method to display key results of the lifelihood program in a concise format.
}
