% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surv}
\alias{surv}
\title{Probability of dying between t and t+dt afer living until t.}
\usage{
surv(t, param1, param2, family = c("exp", "wei", "gam", "lgn"))
}
\arguments{
\item{t}{Numeric. The time to event}

\item{param1}{Numeric. The expected longevity or time to maturity.}

\item{param2}{Numeric. The second parameter returned by lifelihood.}

\item{family}{One of "exp", "wei", "gam", "lgn"}
}
\value{
The probability of event (being alive or not mature) at time t.
}
\description{
Probability of dying between t and t+dt afer living until t.
}
\keyword{internal}
