% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SurvExp}
\alias{SurvExp}
\title{Exponential survival function}
\usage{
SurvExp(t, ExpLong)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{ExpLong}{Numeric. The expected longevity}

\item{Shape}{Numeric. The shape parameter}
}
\value{
The survival probability (numeric)
}
\description{
Exponential survival function
}
\keyword{internal}
