% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SurvGam}
\alias{SurvGam}
\title{Gamma survival function}
\usage{
SurvGam(t, expt_time_to_event, scale)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{expt_time_to_event}{Numeric. The expected longevity or time to maturity.}

\item{scale}{Numeric. The scale parameter}
}
\value{
The survival probability (numeric)
}
\description{
Gamma survival function
}
\keyword{internal}
