% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SurvLgn}
\alias{SurvLgn}
\title{Log-normal survival function}
\usage{
SurvLgn(t, expt_time_to_event, vp1)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{expt_time_to_event}{Numeric. The expected longevity or time to maturity.}

\item{vp1}{Numeric. The vp1 parameter}
}
\value{
The survival probability (numeric)
}
\description{
Log-normal survival function
}
\keyword{internal}
