% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SurvWei}
\alias{SurvWei}
\title{Weibull survival function}
\usage{
SurvWei(t, ExpLong, Shape)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{ExpLong}{Numeric. The expected longevity}

\item{Shape}{Numeric. The shape parameter}
}
\value{
Numeric. The survival probability
}
\description{
Weibull survival function
}
\keyword{internal}
