% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\name{validate_groupby_arg}
\alias{validate_groupby_arg}
\title{Check that the \code{groupby} argument is valid}
\usage{
validate_groupby_arg(lifelihoodData, groupby)
}
\arguments{
\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}.}

\item{groupby}{covariate(s) over which mortality rate should be
computed (default is \code{NULL}). If NULL, calculates a single overall
mortality rate. If \code{"all"}, calculates mortality rate over each
combination of covariates listed in the\code{lifelihoodData} object provided.}
}
\value{
The valid \code{groupby} value
}
\description{
Check that \code{groupby} has an expected value, and returns it
}
\keyword{internal}
