% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_param_range}
\alias{write_param_range}
\title{write parameter ranges/boundaries to a file}
\usage{
write_param_range(data, file_name = "param_range.txt")
}
\arguments{
\item{data}{a dataframe with 3 columns: parameter names, minimum value and maximum value}

\item{file_name}{path of where the .txt file will be stored}
}
\description{
(internal function) \code{write_param_range()} takes a dataframe of parameter ranges/boundaries and write it as .txt file
}
