% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{AIC}
\alias{AIC}
\alias{AIC.lifelihoodResults}
\title{AIC}
\usage{
\method{AIC}{lifelihoodResults}(object)
}
\arguments{
\item{object}{\code{lifelihoodResults} object from \code{\link[=lifelihood]{lifelihood()}}}
}
\value{
The AIC.
}
\description{
S3 method to compute AIC (Akaike information criterion).
}
\seealso{
\code{\link{BIC}}
}
