% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{AICc}
\alias{AICc}
\alias{AICc.lifelihoodResults}
\title{AICc}
\usage{
AICc.lifelihoodResults(object)
}
\arguments{
\item{object}{\code{lifelihoodResults} object from \code{\link[=lifelihood]{lifelihood()}}}
}
\value{
The AICc.
}
\description{
S3 method to compute AICc (Akaike information criterion corrected for small sample size, see Hurvich and Tsai 1989).
}
\seealso{
\code{\link{AIC}}
}
