% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{BIC.lifelihoodResults}
\alias{BIC.lifelihoodResults}
\title{Bayesian Information Criterion}
\usage{
\method{BIC}{lifelihoodResults}(object, ...)
}
\arguments{
\item{object}{output of \code{\link[=lifelihood]{lifelihood()}}}

\item{...}{Ignored}
}
\value{
The BIC
}
\description{
S3 method to compute BIC (Bayesian Information Criterion).
}
\seealso{
\code{\link[=AIC]{AIC()}}, \code{\link[=AICc]{AICc()}}
}
