% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boundaries.R
\name{check_valid_estimation}
\alias{check_valid_estimation}
\title{Check if estimation are too close from boundaries.}
\usage{
check_valid_estimation(results_lifelihood)
}
\description{
Check if estimation are too close from boundaries.
}
\keyword{internal}
