% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{coef}
\alias{coef}
\alias{coef.lifelihoodResults}
\alias{coeff}
\title{Coefficient estimates}
\usage{
\method{coef}{lifelihoodResults}(object, ...)

coeff(object, parameter_name)
}
\arguments{
\item{object}{output of \code{\link[=lifelihood]{lifelihood()}}}

\item{...}{Ignored}

\item{parameter_name}{Name of the parameters to extract the estimate from to extract
all parameter estimates). All parameters#' can be found
\href{/articles/setting-up-the-configuration-file.html#parameters}{here}.}
}
\value{
A nested list of coefficient estimates

A list of coefficient estimates
}
\description{
\code{coef()} retrieve all coefficients from the output of \code{\link[=lifelihood]{lifelihood()}}

\code{coeff()} retrieve coefficients of one parameter
from the output of \code{\link[=lifelihood]{lifelihood()}}
}
