% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{coef}
\alias{coef}
\alias{coef.lifelihoodResults}
\title{Coefficient estimates}
\usage{
\method{coef}{lifelihoodResults}(lifelihoodResults, parameter_name = NULL)
}
\arguments{
\item{lifelihoodResults}{output of \code{\link[=lifelihood]{lifelihood()}}}

\item{parameter_name}{Name of the parameters to extract the estimate from to extract
all parameter estimates). All parameters#' can be found
\href{/articles/setting-up-the-configuration-file.html#parameters}{here}.}
}
\value{
A list of coefficient estimates
}
\description{
S3 method to retrieve coefficients from the output of \code{\link[=lifelihood]{lifelihood()}}
}
\examples{
df <- lifelihood::fakesample
df$type <- as.factor(df$type)
df$geno <- as.factor(df$geno)

clutchs <- c(
  "clutch_start1", "clutch_end1", "clutch_size1",
  "clutch_start2", "clutch_end2", "clutch_size2"
)

dataLFH <- lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "death_start",
  death_end = "death_end",
  covariates = c("geno", "type"),
  model_specs = c("gam", "lgn", "wei")
)

results <- lifelihood(
  lifelihoodData = dataLFH,
  path_config = here::here("config.yaml"),
  seeds = c(1, 2, 3, 4),
  raise_estimation_warning = FALSE
)
coef(results)
coef(results, "expt_death")
}
