% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_event_rate.R
\name{compute_fitted_event_rate}
\alias{compute_fitted_event_rate}
\title{Compute fitted mortality rate}
\usage{
compute_fitted_event_rate(
  lifelihoodResults,
  interval_width,
  event = c("mortality", "maturity", "reproduction"),
  newdata = NULL,
  max_time = NULL,
  groupby = NULL,
  mcmc.ci.fit = FALSE
)
}
\arguments{
\item{lifelihoodResults}{output of \code{\link[=lifelihood]{lifelihood()}}.}

\item{interval_width}{The interval width used to calculate the
event rate. For instance, if the time unit for deaths in
the original dataset is days and \code{interval_width} is set to 10,
the event rate will be calculated every 10 days for each group.}

\item{event}{Which event to compute? Must be one of "mortality", "maturity", "reproduction".}

\item{newdata}{Optional \code{data.frame} providing covariate values for prediction.
If \code{NULL}, the original model data are used.}

\item{max_time}{The maximum time for calculating the event
rate. If set to NULL, the time of the last observed death is used.}

\item{groupby}{One or multiple covariates used to group the computation.}

\item{mcmc.ci.fit}{Whether or not to retrieve MCMC CI estimations.}
}
\value{
A dataframe with 3 columns: Interval (time interval, based
on \code{interval_width} value), group (identifier of a given subgroup,
or "Overall" if groupby = NULL), and Event_rate (event rate
over the interval).
Note that for reproduction event, the first reproduction event of
each individual cannot be computed if maturity was not observed (i.e. mat_clutch is true)
When the interval between the last reproduction event of an individual
and their death is greater than \code{interval_width} the individuals are
included in the computation of reproduction rate
}
\description{
Calculate the empirical mortality rate over a given interval
on some new data.
}
\examples{
df <- datapierrick |>
tibble::as_tibble() |>
 dplyr::mutate(par = as.factor(par))

# name of the columns of the clutchs into a single vector
clutchs <- generate_clutch_vector(28)
dataLFH <- as_lifelihoodData(
 df = df,
 sex = "sex",
 sex_start = "sex_start",
 sex_end = "sex_end",
 maturity_start = "mat_start",
 maturity_end = "mat_end",
 clutchs = clutchs,
 death_start = "death_start",
 death_end = "death_end",
 matclutch = FALSE,
 covariates = c("par", "geno"),
 dist = c(mortality = "wei", maturity = "gam", reproduction = "lgn")
)
results <- lifelihood(
lifelihoodData = dataLFH,
config = list(
 mortality = list(expt_death = "par", survival_param2 = 1),
 maturity = list(expt_maturity = 1, maturity_param2 = 1),
 reproduction = list(
   expt_reproduction = 1,
   reproduction_param2 = 1,
   n_offspring = 1
 )
),
seeds = c(1, 2, 3, 4)
)
fitted_emergence_rate <- compute_fitted_event_rate(
 lifelihoodResults = results,
 interval_width = 5,
 event = c("mortality"),
 max_time=150,
 groupby=c("par"))|>
 dplyr::mutate(sex = "sex=all")

p <- fitted_emergence_rate |>
 ggplot2::ggplot(
   ggplot2::aes(
     x = Mean_Interval,
     y = Event_Rate,
     color = par,
     shape = par
   )
 )+
 ggplot2::geom_point()+
 ggplot2::geom_line(linewidth=0.5)+
 ggplot2::xlab("Time (days)")+
 ggplot2::ylab("Fitted mortality rate over 5 day-periods")+
 ggplot2::facet_wrap(ggplot2::vars(sex, par), labeller = "label_both")
p

}
