% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_event_rate.R
\name{compute_fitted_event_rate}
\alias{compute_fitted_event_rate}
\title{Compute fitted mortality rate}
\usage{
compute_fitted_event_rate(
  lifelihoodResults,
  interval_width,
  event = c("mortality", "maturity", "reproduction"),
  newdata = NULL,
  max_time = NULL,
  groupby = NULL
)
}
\arguments{
\item{lifelihoodResults}{output of \code{\link[=lifelihood]{lifelihood()}}.}

\item{interval_width}{The interval width used to calculate the
event rate. For instance, if the time unit for deaths in
the original dataset is days and \code{interval_width} is set to 10,
the event rate will be calculated every 10 days for each group.}

\item{event}{Which event to compute? Must be one of "mortality", "maturity", "reproduction".}

\item{newdata}{Optional \code{data.frame} providing covariate values for prediction.
If \code{NULL}, the original model data are used.}

\item{max_time}{The maximum time for calculating the event
rate. If set to NULL, the time of the last observed death is used.}

\item{groupby}{One or multiple covariates used to group the computation.}
}
\value{
A dataframe with 3 columns: Interval (time interval, based
on \code{interval_width} value), group (identifier of a given subgroup,
or "Overall" if groupby = NULL), and Event_rate (event rate
at this time).
}
\description{
Calculate the empirical mortality rate over a given interval
on some new data.
}
