% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compute_mortality_rate}
\alias{compute_mortality_rate}
\title{Compute empirical mortality rate}
\usage{
compute_mortality_rate(
  data,
  start_col,
  end_col,
  covariates,
  interval_width = 2,
  max_time = NULL
)
}
\arguments{
\item{data}{Dataframe with the data of life history. It should have one row per life history / observation.}

\item{start_col}{Column name containing the first date of the interval in which the event was determined.}

\item{end_col}{Column name containing the second date of the interval in which the event was determined.}

\item{covariates}{Vector containing the names of the covariates.}

\item{interval_width}{Width of the interval to compute the mortality rate.}

\item{max_time}{Maximum time to compute the mortality rate.}
}
\value{
none
}
\description{
Compute the empirical mortality rate
}
