% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_boundaries.R
\name{create_default_boundaries}
\alias{create_default_boundaries}
\title{Dataframe with default boundaries}
\usage{
create_default_boundaries(model_specs, max_death, max_clutch, max_maturity)
}
\description{
Dataframe with default boundaries
}
