% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_boundaries.R
\name{default_bounds_df}
\alias{default_bounds_df}
\title{Dataframe with default boundaries}
\usage{
default_bounds_df(lifelihoodData)
}
\arguments{
\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}}
}
\description{
Dataframe with default boundaries
}
