% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delink}
\alias{delink}
\title{Delink function}
\usage{
delink(obs, min_and_max)
}
\arguments{
\item{obs}{Numeric. The observed value of the parameter}

\item{min_and_max}{Numeric vector of length 2. The minimum and maximum values of the parameter}
}
\value{
Numeric. The transformed parameter
}
\description{
Delink function to transform the parameters from original scale to lifelihood scale.
}
\keyword{internal}
