% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_os}
\alias{detect_os}
\title{Group of internal functions used in the lifelihood programme}
\usage{
detect_os()
}
\value{
string of operatin system
}
\description{
\code{detect_os()} finds the OS name
}
