% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_os}
\alias{detect_os}
\title{Find the operating system of the user}
\usage{
detect_os()
}
\value{
String with the name of the operating system, either "Windows" or "Unix-like"
}
\description{
\code{detect_os()} finds the operating system name
}
\examples{
detect_os()

}
\keyword{internal}
