% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_bin.R
\name{execute_bin}
\alias{execute_bin}
\title{Execution of the compiled files}
\usage{
execute_bin(
  path_to_Lifelihood,
  path_input_data,
  path_param_bounds,
  MCMC,
  interval,
  se.fit,
  saveprobevent,
  fitness,
  r,
  seed1,
  seed2,
  seed3,
  seed4,
  ratiomax,
  tc,
  tinf,
  sub_interval,
  path_continuous_var,
  ntr,
  nst,
  To,
  Tf,
  climbrate,
  precision
)
}
\arguments{
\item{path_to_Lifelihood}{A character string specifying the file
path to the compile Lifelihood program (default is NULL).}

\item{path_input_data}{Path to the input text file with the model
and data to be fitted.}

\item{path_param_bounds}{Path to the parameter boundaries text file
with the min and max boudaries for each parameter.}

\item{MCMC}{Perform MCMC sampling of the parameter after convergence
to estimate their 95\% confidence interval.}

\item{interval}{TBD - Check the actual meaning.}

\item{se.fit}{If TRUE, Lifelihood computes the standard error of
each parameter using the variance-covariance matrix.}

\item{saveprobevent}{TBD - Check the actual meaning.}

\item{fitness}{Reparametrize the model with one parameter as the
intrinsic rate of increase.}

\item{r}{Reparametrize the model with one parameter as the intrinsic
rate of increase.}

\item{seed1}{First seed number used to reproduce results (same seed
= same results).}

\item{seed2}{Second seed number used to reproduce results (same seed
= same results).}

\item{seed3}{Third seed number used to reproduce results (same seed
= same results).}

\item{seed4}{Fourth seed number used to reproduce results (same seed
= same results).}

\item{ratiomax}{Maximum multiplicative factor for clutch size in
models with reproductive senescence.}

\item{tc}{Critical age (after the juvenile mortality peak) at which
the survival model starts to be fitted.}

\item{tinf}{Maximum censoring time (should be greater than the age
of the oldest individual observed dead in the dataset).}

\item{sub_interval}{Sub-interval used to integrate the left and
right censoring dates of each event.}

\item{path_continuous_var}{Path to the continuous variables file.}

\item{ntr}{Number of thread for the paralelisation.}

\item{nst}{TBD - Check the actual meaning.}

\item{To}{Initial temperature for the simulated annealing.}

\item{Tf}{Final temperature for the simulated annealing.}

\item{climbrate}{Rate for the simulated annealing.}

\item{precision}{Precision parameter for the algorithm.}
}
\description{
Run lifelihood program in console mode.
}
\keyword{internal}
