% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_to_num}
\alias{factor_to_num}
\title{Convert a factor variable to integers}
\usage{
factor_to_num(x)
}
\arguments{
\item{x}{A column name}
}
\value{
The integer
}
\description{
Function necessary because the Rust program
expects integers for factor levels.
}
\keyword{internal}
