% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_file.R
\name{format_config}
\alias{format_config}
\title{(internal function) Read the configuration file and format it for the input file.}
\usage{
format_config(path_config, covariates)
}
\arguments{
\item{path_config}{Location of the configuration file.}

\item{covariates}{Vector containing the names of the covariates.}
}
\value{
Vector of string to add in the input file
}
\description{
Use the configuration file to create the well formatted vector of strings that will be put under "\emph{\strong{modele}}" in the input file, excluding the distribution law.
}
\keyword{internal}
