% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{format_dataframe_to_txt}
\alias{format_dataframe_to_txt}
\title{create the input data file from a dataframe}
\usage{
format_dataframe_to_txt(df, extra1 = NULL, extra2 = NULL, extra3 = NULL)
}
\arguments{
\item{df}{the dataframe of the input data}

\item{extra1}{name of the first column to add in the input data file}

\item{extra2}{name of the second column to add in the input data file}

\item{extra3}{name of the third column to add in the input data file}

\item{output_file}{path used to write the input data file}
}
\description{
takes a dataframe and apply to each row the \code{\link[=format_row]{format_row()}} function to create the input data file
}
\details{
the number of extra column is \strong{currently limited} to 3 for simplicity, but we definitly should \strong{change} this accept an unlimited number of extra columns
}
