% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{format_dataframe_to_txt}
\alias{format_dataframe_to_txt}
\title{Create the input data file from a dataframe}
\usage{
format_dataframe_to_txt(
  df,
  sex,
  sex_start,
  sex_end,
  maturity_start,
  maturity_end,
  matclutch,
  matclutch_size,
  clutchs,
  death_start,
  death_end,
  covariates,
  model_specs,
  path_config,
  temp_dir
)
}
\arguments{
\item{df}{The dataframe object of the input data.}

\item{sex}{Column name containing the sex of the observations.}

\item{sex_start}{Column name containing the first date of the interval in which the sex was determined.}

\item{sex_end}{Column name containing the second date of the interval in which the sex was determined.}

\item{maturity_start}{Column name containing the first date of the interval in which the maturity was determined.}

\item{maturity_end}{Column name containing the second date of the interval in which the maturity was determined.}

\item{matclutch}{Whether the maturity event (designated by \code{maturity_start} and \code{maturity_end}) is a clutch event or not. If \code{TRUE}, must specify the \code{matclutch_size} argument.}

\item{matclutch_size}{Column name containing the size of the clutch for the maturity event. Only used (and required) if \code{matclutch} is \code{TRUE}.}

\item{clutchs}{Vector containing the names of the clutch columns. The order should be:
\itemize{
\item first date of the first clutch
\item second date of first clutch
\item clutch size of the first clutch
\item second clutch first date
\item first clutch second date
\item second clutch clutch size
\item and so on.
If the observation with the most clutches is for example 10, then the vector must be of size 10 x 3 = 30 (3 elements per clutch: first date, second date and size).
}}

\item{death_start}{Column name containing the first date of the interval in which the death was determined.}

\item{death_end}{Column name containing the second date of the interval in which the death was determined.}

\item{covariates}{Vector containing the names of the covariates.}

\item{model_specs}{Vector of characters with the name of the statistical law to use. Must be of length 3 and each element must be in "wei", "gam" or "lgn". The first one is used for maturity, the second one is used for clutchs and the third one for death.}

\item{path_config}{Path to the configuration file (YAML).}

\item{temp_dir}{Name of the temporary directory with temporary files}
}
\description{
Takes a dataframe (\code{df} argument in \code{\link[=lifelihood]{lifelihood()}} function) and apply to each row the \code{\link[=format_row]{format_row()}} function to create the input data file.
}
\keyword{internal}
