% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{format_row}
\alias{format_row}
\title{Format a dataframe row for the input data file}
\usage{
format_row(
  row,
  sex,
  sex_start,
  sex_end,
  maturity_start,
  maturity_end,
  clutchs,
  death_start,
  death_end,
  covariates
)
}
\arguments{
\item{row}{A row of the dataframe object provided by the user (\code{df} argument in \code{\link[=lifelihood]{lifelihood()}} function).}
}
\value{
A string of the well formated row.
}
\description{
(internal function) Takes a row from a dataframe with input data (sex, maturity, clutch date and death) in interval format and transform it into a large string (required format for the data input file).
}
