% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{format_row}
\alias{format_row}
\title{Format a dataframe row for the input data file}
\usage{
format_row(
  row,
  sex,
  sex_start,
  sex_end,
  maturity_start,
  maturity_end,
  clutchs,
  death_start,
  death_end,
  covariates
)
}
\arguments{
\item{row}{A row of the dataframe object provided by the user (\code{df} argument in \code{\link[=lifelihood]{lifelihood()}} function).}

\item{sex}{Column name containing the sex of the observations.}

\item{sex_start}{Column name containing the first date of the interval in which the sex was determined.}

\item{sex_end}{Column name containing the second date of the interval in which the sex was determined.}

\item{maturity_start}{Column name containing the first date of the interval in which the maturity was determined.}

\item{maturity_end}{Column name containing the second date of the interval in which the maturity was determined.}

\item{clutchs}{Vector containing the names of the clutch columns. The order should be:
\itemize{
\item first date of the first clutch
\item second date of first clutch
\item clutch size of the first clutch
\item second clutch first date
\item first clutch second date
\item second clutch clutch size
\item and so on.
}

If the observation with the most clutches is for example 10, then the vector must be of size 10 x 3 = 30 (3 elements per clutch: first date, second date and size).}

\item{death_start}{Column name containing the first date of the interval in which the death was determined.}

\item{death_end}{Column name containing the second date of the interval in which the death was determined.}

\item{covariates}{Vector containing the names of the covariates.}

\item{matclutch}{Whether the maturity event (designated by \code{maturity_start} and \code{maturity_end}) is a clutch event or not. If \code{TRUE}, must specify the \code{matclutch_size} argument.}

\item{matclutch_size}{Column name containing the size of the clutch for the maturity event. Only used (and required) if \code{matclutch} is \code{TRUE}.}
}
\value{
A string of the well formated row.
}
\description{
Takes a row from a dataframe with input data (sex, maturity, clutch date and death) in interval format and transform it into a large string (required format for the data input file).
}
\keyword{internal}
