% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{format_row}
\alias{format_row}
\title{Format a dataframe row for the input data file}
\usage{
format_row(
  row,
  sex,
  sex_start,
  sex_end,
  maturity_start,
  maturity_end,
  clutchs,
  death_start,
  death_end,
  covariates
)
}
\arguments{
\item{row}{A row of the dataframe object provided by the user
(\code{df} argument in \code{\link[=lifelihoodData]{lifelihoodData()}} function).}

\item{sex}{Column name containing the sex of the observations.}

\item{sex_start}{Column name containing the first date of the
interval in which the sex was determined.}

\item{sex_end}{Column name containing the second date of the
interval in which the sex was determined.}

\item{maturity_start}{Column name containing the first date of
the interval in which the maturity was determined.}

\item{maturity_end}{Column name containing the second date of
the interval in which the maturity was determined.}

\item{clutchs}{Vector containing the names of the clutch columns.}

\item{death_start}{Column name containing the first date of the
interval in which the death was determined.}

\item{death_end}{Column name containing the second date of the
interval in which the death was determined.}

\item{covariates}{Vector containing the names of the covariates.}
}
\value{
A string of the well formated row.
}
\description{
Takes a row from a dataframe with input data (sex, maturity,
clutch date and death) in interval format and transform it
into a large string (required format for the data input file).
}
\keyword{internal}
