% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_test_config}
\alias{use_test_config}
\title{Get the path to a built-in configuration file.}
\usage{
use_test_config(config_name = c("config", "config2", "config_pierrick"))
}
\arguments{
\item{config_name}{Configuration name. Currently available options:
\itemize{
\item config
\item config2
\item config_pierrick
By default, it will use "config".
}}
}
\value{
Absolute path to the configuration file
}
\description{
\code{lifelihood} embeds a few configuration files, and
this function is a simple tool to access one of them.

It takes the name one of the available configuration
and returns the path to it.

For more info about configuration files, see
\code{vignette("setting-up-the-configuration-file", package = "lifelihood")}
}
\examples{
use_test_config("config")
use_test_config("config2")

}
