% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_param_ranges}
\alias{get_param_ranges}
\title{get the parameter ranges/boundaries in the output file}
\usage{
get_param_ranges(lines)
}
\arguments{
\item{lines}{vector of the output file, where each element is a line of the file}
}
\description{
find the parameter ranges/boundaries in the output file of the lifelihood program
}
