% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_seeds}
\alias{get_seeds}
\alias{get_likelihood}
\alias{get_param_ranges}
\alias{get_effects}
\alias{get_ratio_max}
\alias{parse}
\title{Data Parsing Functions for Model Output Files}
\usage{
get_seeds(lines, group_by_group = FALSE)

get_likelihood(lines, group_by_group = FALSE)

get_param_ranges(lines)

get_effects(lines, group_by_group = FALSE)

get_ratio_max(lines)

parse(lines, element, group_by_group = FALSE)
}
\arguments{
\item{lines}{A character vector where each element represents a line from the output file.}

\item{group_by_group}{A logical indicating whether to parse the data in a group-by-group manner. Default is FALSE.}
}
\value{
Depending on the function:
\itemize{
\item \code{get_seeds}: Returns a numeric vector of seeds or a matrix of seeds if \code{group_by_group} is TRUE.
\item \code{get_likelihood}: Returns a numeric value of the likelihood or a matrix of likelihood values if \code{group_by_group} is TRUE.
\item \code{get_param_ranges}: Returns a data frame with parameter names and their respective min and max values.
\item \code{get_effects}: Returns a data frame with effect names, estimates, and standard errors.
\item \code{get_ratio_max}: Returns a numeric value of the ratio maximum.
}
}
\description{
These functions parse specific elements from model output files, including seeds, likelihood values, parameter ranges, effects, and ratio maximum values. The functions can handle group-by-group parsing where applicable.
}
\examples{

lines <- c(
  "seed1= 123 seed2= 456 seed3= 789 seed4= 101",
  "Likelihood_max= -123.456",
  "Parameter_Range_Table",
  "param1 0 10",
  "param2 5 15",
  "ratiomax 20",
  "eff_param1 0.5 0.1",
  "eff_param2 0.8 0.2"

seeds <- get_seeds(readLines(file_path1))
seeds_gbg <- get_seeds(readLines(file_path2), group_by_group=TRUE)
likelihood <- get_likelihood(readLines(file_path1))
likelihood_gbg <- get_likelihood(readLines(file_path2), group_by_group=TRUE)
param_ranges <- get_param_ranges(readLines(file_path1))
effects <- get_effects(readLines(file_path1))
ratio_max <- get_ratio_max(readLines(file_path1))


}
