% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{get_seeds}
\alias{get_seeds}
\title{get the seeds in the output file}
\usage{
get_seeds(lines, group_by_group = FALSE)
}
\arguments{
\item{lines}{vector of the output file, where each element is a line of the file}

\item{group_by_group}{boolean indicating whether the analysis should be made group by group or not (default to false)}
}
\description{
find the seeds value in the output file of the lifelihood program
}
\details{
List of parsing functions use to read the output file of the lifelihood program
}
