% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{lifelihood}
\alias{lifelihood}
\title{Individual life history modelling}
\usage{
lifelihood(
  lifelihoodData,
  path_config,
  path_to_Lifelihood = NULL,
  param_bounds_df = NULL,
  group_by_group = FALSE,
  MCMC = 0,
  interval = 25,
  SEcal = FALSE,
  saveprobevent = 0,
  fitness = 0,
  r = 0,
  seeds = NULL,
  ntr = 2,
  nst = 2,
  To = 50,
  Tf = 1,
  climbrate = 1,
  precision = 0.001,
  raise_estimation_warning = TRUE,
  delete_temp_files = TRUE
)
}
\arguments{
\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}.}

\item{path_config}{A character string specifying the file path to the YAML configuration file.}

\item{path_to_Lifelihood}{A character string specifying the file path to the compile Lifelihood program (default is NULL)}

\item{param_bounds_df}{Dataframe with the parameter ranges/boundaries/boundaries}

\item{group_by_group}{Boolean option to fit the full factorail model with all the interactions between each of the factors}

\item{MCMC}{Perform MCMC sampling of the parameter after convergence to estimate their 95\% confidence interval}

\item{interval}{TBD - Check the actual meaning}

\item{SEcal}{If TRUE, Lifelihood computes the standard error of each parameter using the Hessian matrix (output with value of -1 if standard error cannot be computed due to singularity of the Hessian matrix)}

\item{saveprobevent}{TBD - Check the actual meaning}

\item{fitness}{Reparametrize the model with one parameter as the lifetime reproductive success}

\item{r}{Reparametrize the model with one parameter as the intrinsic rate of increase}

\item{seeds}{Numbers used to reproduce results (same seeds = same results). This must be a vector of length 4.}

\item{ntr}{Number of thread for the paralelisation ?}

\item{nst}{TBD - Check the actual meaning}

\item{To}{Initial temperature for the simulated annealing}

\item{Tf}{Initial temperature for the simulated annealing}

\item{climbrate}{Rate for the simulated annealing ?}

\item{precision}{TBD - Check the actual meaning}

\item{raise_estimation_warning}{Whether or not to raise a warning when the estimate of a parameter is too close to its minimum or maximum bound. Default is TRUE.}

\item{delete_temp_files}{Indicates whether temporary files should be deleted. TRUE by default and recommended.}
}
\value{
A \code{lifelihoodResults} object
}
\description{
Computes the joined likelihood of all the events in
an individual life-history (time of maturity, reproductive
events, death) and estimates the parameters of the model
using maximum likelihood.
}
\examples{
library(lifelihood)

df <- fakesample
head(df)
df$type <- as.factor(df$type)
df$geno <- as.factor(df$geno)

clutchs <- c(
  "clutch_start1", "clutch_end1", "clutch_size1",
  "clutch_start2", "clutch_end2", "clutch_size2"
)

dataLFH <- lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "death_start",
  death_end = "death_end",
  covariates = c("geno", "type"),
  model_specs = c("gam", "lgn", "wei")
)

results <- lifelihood(
  lifelihoodData = dataLFH,
  path_config = get_config_path("config"),
  seeds = c(1, 2, 3, 4),
  raise_estimation_warning = FALSE
)
summary(results)
}
