% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_group.R
\name{lifelihood_fit_group_by_group}
\alias{lifelihood_fit_group_by_group}
\title{Fit lifelihood group by group}
\usage{
lifelihood_fit_group_by_group(
  lifelihoodData,
  path_config,
  path_to_Lifelihood = NULL,
  param_bounds_df = NULL,
  MCMC = 0,
  interval = 25,
  se.fit = FALSE,
  saveprobevent = 0,
  r = 0,
  seeds = NULL,
  ntr = 2,
  nst = 2,
  To = 50,
  Tf = 1,
  climbrate = 1,
  precision = 0.001,
  ratiomax = 10,
  tc = 20,
  tinf = 1000,
  sub_interval = 0.3,
  delete_temp_files = TRUE
)
}
\arguments{
\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=as_lifelihoodData]{as_lifelihoodData()}}.}

\item{path_config}{A character string specifying the file path to the YAML configuration file.}

\item{path_to_Lifelihood}{A character string specifying the file path to the compile Lifelihood program (default is NULL)}

\item{param_bounds_df}{Dataframe with the parameter ranges/boundaries/boundaries}

\item{MCMC}{Perform MCMC sampling of the parameter after convergence to estimate their 95\% confidence interval}

\item{interval}{TBD - Check the actual meaning}

\item{se.fit}{If TRUE, Lifelihood computes the standard error of each parameter using the variance-covariance matrix (approximated by the inverse of the negative of the Hessian matrix).}

\item{saveprobevent}{TBD - Check the actual meaning}

\item{r}{Reparametrize the model with one parameter as the intrinsic rate of increase}

\item{seeds}{Vector of length for with seed numbers used to reproduce results (same seeds = same results). This argument should be \code{NULL} (default) when \code{n_fit} > 1.}

\item{ntr}{Number of thread for the paralelisation ?}

\item{nst}{TBD - Check the actual meaning}

\item{To}{Initial temperature for the simulated annealing}

\item{Tf}{Initial temperature for the simulated annealing}

\item{climbrate}{Rate for the simulated annealing ?}

\item{precision}{TBD - Check the actual meaning}

\item{ratiomax}{Maximum multiplicative factor for clutch size in models with reproductive senescence (cf CalculRatioEspPoissonTronque function in Lifelihood)}

\item{tc}{critical age (after the juvenile mortality peak) at which the survival model starts to be fitted}

\item{tinf}{maximum censoring time (should be greater than the age of the oldest individual observed dead in the dataset)}

\item{sub_interval}{sub-interval used to integrate the left and right censoring dates of each event}

\item{delete_temp_files}{Indicates whether temporary files should be deleted. TRUE by default and recommended.}
}
\value{
A \code{lifelihoodResults} object with group-by-group results.
}
\description{
Orchestrator function that splits data by covariate groups,
fits intercept-only models for each group, and merges results.
}
\keyword{internal}
