% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Link function}
\usage{
link(estimate, min, max)
}
\arguments{
\item{estimate}{Numeric. The estimate of the parameter}

\item{min_and_max}{Numeric vector of length 2. The minimum and maximum values of the parameter}
}
\value{
Numeric. The transformed parameter
}
\description{
Link function to transform the parameters from the lifelihood scale to the original scale
}
\keyword{internal}
