% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{Compute and visualize predicted mortality rate}
\usage{
make_design_matrix(covariates, data)
}
\description{
Create the design matrix for the expected death and the survival shape.
}
\keyword{internal}
