% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_output}
\alias{parse_output}
\alias{get_seeds}
\alias{get_likelihood}
\alias{get_param_ranges}
\alias{get_ratio_max}
\alias{get_effects}
\alias{get_hessian}
\alias{get_mcmc}
\title{Parsing functions used to read the output file of the program}
\usage{
parse_output(lines, element)

get_seeds(lines)

get_likelihood(lines)

get_param_ranges(lines)

get_ratio_max(lines)

get_effects(lines)

get_hessian(lines)

get_mcmc(lines)
}
\arguments{
\item{lines}{Vector of the output file (\code{.out}), where each element is a line of the file.}

\item{element}{Name of the result to parse. Must be in one of 'seeds', 'likelihood', 'effects', 'parameter_ranges', 'ratio_max', 'mcmc'.}
}
\value{
The parsed element
}
\description{
Find specific result in the output file of the lifelihood
program, according to the \code{element} argument. This function
is an aggregator of all the \verb{get_*()} functions described below.
}
\keyword{internal}
