% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{plot.lifelihoodGOF}
\alias{plot.lifelihoodGOF}
\title{Plot goodness-of-fit histogram}
\usage{
\method{plot}{lifelihoodGOF}(gof, bins = 30, fill = "grey80", color = "white", line_color = "red", ...)
}
\arguments{
\item{bins}{Number of histogram bins.}

\item{fill}{Histogram fill color.}

\item{color}{Histogram border color.}

\item{line_color}{Color of the vertical line showing original log-likelihood.}

\item{...}{Ignored.}

\item{x}{Output of \code{\link[=goodness_of_fit]{goodness_of_fit()}}.}
}
\value{
A ggplot object.
}
\description{
Plot a histogram of simulated/refitted log-likelihood values and overlay a
vertical line for the original model log-likelihood.
}
