% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_event_rate}
\alias{plot_event_rate}
\alias{plot_fitted_event_rate}
\alias{plot_observed_event_rate}
\title{Display change in predicted event rate with age}
\usage{
plot_fitted_event_rate(
  lifelihoodResults,
  interval_width,
  event = c("mortality", "maturity", "reproduction"),
  newdata = NULL,
  add_observed_event_rate = TRUE,
  min_sample_size = 1,
  max_time = NULL,
  groupby = NULL,
  use_facet = FALSE,
  xlab = "Time",
  ylab = "Event Rate"
)

plot_observed_event_rate(
  lifelihoodData,
  interval_width,
  event = c("mortality", "maturity", "reproduction"),
  max_time = NULL,
  min_sample_size = 1,
  groupby = NULL,
  use_facet = FALSE,
  xlab = "Time",
  ylab = "Mortality Rate"
)

plot_event_rate(
  rate_df,
  max_time,
  type = c("points", "lines"),
  groupby,
  use_facet,
  xlab = "Time",
  ylab = "Event rate"
)
}
\arguments{
\item{interval_width}{The interval width used to calculate the
event rate. For instance, if the time unit for deaths in
the original dataset is days and \code{interval_width} is set to 10,
the event rate will be calculated every 10 days for each group.}

\item{event}{Which event to compute? Must be one of "mortality", "maturity", "reproduction".}

\item{newdata}{Data for prediction. If absent, predictions are for the subjects used in the original fit.}

\item{add_observed_event_rate}{Boolean to add the observed event rate to the graph (default=TRUE)}

\item{min_sample_size}{The minimum number of individuals alive
at the beggining of a time interval for computing the observed event rate (only used if add_observed_event_rate=TRUE, default=1)}

\item{max_time}{The maximum time for calculating the event
rate. If set to NULL, the time of the last observed death is used.}

\item{groupby}{Factor(s) whosse levels over which event rate should be represented (default=NULL)}

\item{use_facet}{Use facet_wrap to plot one panel per group (default=FALSE)}

\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}.}

\item{rate_df}{Dataframe with event rate}

\item{type}{The type of symbol to be used for the plot (either of "points" or 'lines")}
}
\value{
a ggplot2 plot

a ggplot2 plot

a ggplot2 plot
}
\description{
Useful function for creating a good-quality line graph
of changes in predicted event rate with age.

If you want more control over the style of the graph,
use the \code{\link[=compute_fitted_event_rate]{compute_fitted_event_rate()}} function to directly retrieve the predicted data.

Useful function for creating a good-quality line graph
of changes in the empirical mortality rate.

You don't need to fit with \code{\link[=lifelihood]{lifelihood()}} to use this
function, only to retrieve a lifelihood data object
with \code{\link[=lifelihoodData]{lifelihoodData()}}

If you want more control over the style of the graph,
use the \code{\link[=compute_observed_event_rate]{compute_observed_event_rate()}} function to retrieve data.

Convenient function used in \code{\link[=plot_observed_event_rate]{plot_observed_event_rate()}}
and \code{\link[=plot_fitted_event_rate]{plot_fitted_event_rate()}}.
}
\details{
This function requires \href{https://ggplot2.tidyverse.org/}{ggplot2} to be installed.

This function requires \href{https://ggplot2.tidyverse.org/}{ggplot2} to be installed.
}
\keyword{internal}
