% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mortality_rate}
\alias{plot_mortality_rate}
\title{Display evolution of empirical mortality rate}
\usage{
plot_mortality_rate(
  interval_width,
  lifelihoodData = NULL,
  lifelihoodResults = NULL,
  prediction = FALSE,
  newdata = NULL,
  max_time = NULL,
  groupby = NULL,
  log_x = FALSE,
  log_y = FALSE,
  plot_title = "Mortality rate over time",
  xlab = "Time",
  ylab = "Mortality Rate"
)
}
\arguments{
\item{interval_width}{The interval width used to calculate the mortality rate. For instance, if the time unit for deaths in the original dataset is days and \code{interval_width} is set to 10, the mortality rate will be calculated every 10 days for each group.}

\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}.}

\item{lifelihoodResults}{Output object of the \code{\link[=lifelihood]{lifelihood()}} function.}

\item{prediction}{Decide whether the plotted mortality rate should be based on
empirical data using \code{\link[=lifelihoodData]{lifelihoodData()}} or on predicted values
generated by \code{\link[=lifelihoodResults]{lifelihoodResults()}} (and \code{\link[=predict]{predict()}}, under the
bonnet). If \code{newdata} is \code{NULL}, predictions are made for the
subjects included in the original model fit.}

\item{newdata}{Data for prediction. If absent, predictions are for the subjects used in the original fit.}

\item{max_time}{The maximum time for calculating the mortality rate. If set to NULL, the time of the last observed death is used.}

\item{groupby}{covariate(s) over which mortality rate should be computed (default is NULL). If NULL, calculates a single overall mortality rate. If "all", calculates mortality rate over each combination of covariates listed in the\code{lifelihoodData} object provided.}

\item{log_x}{Determine whether the x-axis should be displayed on a logarithmic scale.}

\item{log_y}{Determine whether the y-axis should be displayed on a logarithmic scale.}
}
\value{
a ggplot2 plot
}
\description{
Useful function for creating a good-quality line graph
of changes in the empirical mortality rate.
}
\details{
This function requires \href{https://ggplot2.tidyverse.org/}{ggplot2} to be installed.
}
\examples{
library(lifelihood)

df <- datapierrick
df$geno <- as.factor(df$geno)
df$par <- as.factor(df$par)
df$spore <- as.factor(df$spore)
clutchs <- paste("pon", rep(c("start", "end", "size"), 28), rep(1:28, each = 3), sep = "_")

dataLFH <- lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "death_start",
  death_end = "death_end",
  covariates = c("par", "geno", "spore"),
  model_specs = c("gam", "lgn", "wei")
)

plot_mortality_rate(dataLFH, interval_width = 20, max_time = 170)
plot_mortality_rate(
  dataLFH,
  interval_width = 10,
  max_time = 170,
  groupby = NULL,
  log_y = TRUE
)
}
