% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\name{plot_mortality_rate}
\alias{plot_mortality_rate}
\title{Compute and visualize predicted mortality rate}
\usage{
plot_mortality_rate(
  lifelihoodResults,
  newdata,
  intervals,
  use_log_x = FALSE,
  use_log_y = FALSE
)
}
\arguments{
\item{lifelihoodResults}{Output object of the \code{\link[=lifelihood]{lifelihood()}} function.}

\item{newdata}{Dataframe with the values of the covariates to predict the mortality rate}

\item{intervals}{Vector of the intervals (x-axis, same unit as the data) to compute the mortality rate}

\item{use_log_x}{Boolean to use log scale for the x-axis (default: FALSE)}

\item{use_log_y}{Boolean to use log scale for the y-axis (default: FALSE)}
}
\description{
Simple plot of the predicted mortality rate. To access the values, use the output of \code{\link[=pred_mortality_rate]{pred_mortality_rate()}}.
}
