% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mortality_rate}
\alias{plot_mortality_rate}
\alias{plot_fitted_mortality_rate}
\alias{plot_observed_mortality_rate}
\title{Display evolution of predicted mortality rate}
\usage{
plot_fitted_mortality_rate(
  lifelihoodResults,
  interval_width,
  newdata = NULL,
  max_time = NULL,
  groupby = NULL,
  log_x = FALSE,
  log_y = FALSE,
  title = "Mortality rate over time",
  xlab = "Time",
  ylab = "Mortality Rate"
)

plot_observed_mortality_rate(
  lifelihoodData,
  interval_width,
  max_time = NULL,
  min_sample_size = 1,
  groupby = NULL,
  log_x = FALSE,
  log_y = FALSE,
  title = "Mortality rate over time",
  xlab = "Time",
  ylab = "Mortality Rate"
)

plot_mortality_rate(
  rate_df,
  max_time,
  groupby,
  use_facet = FALSE,
  log_x,
  log_y,
  title,
  xlab = "Time",
  ylab = "Mortality rate"
)
}
\arguments{
\item{lifelihoodResults}{output of \code{\link[=lifelihood]{lifelihood()}}}

\item{interval_width}{The interval width used to calculate the
mortality rate. For instance, if the time unit for deaths in
the original dataset is days and \code{interval_width} is set to 10,
the mortality rate will be calculated every 10 days for each group.}

\item{newdata}{Data for prediction. If absent, predictions are for the subjects used in the original fit.}

\item{max_time}{The maximum time for calculating the mortality
rate. If set to NULL, the time of the last observed death is used.}

\item{groupby}{covariate(s) over which mortality rate should be
computed (default is \code{NULL}).
\itemize{
\item If NULL, calculates a single overall mortality rate.
\item If \code{"all"}, calculates mortality rate over each combination
of covariates listed in the\code{lifelihoodData} object provided.
\item Otherwise must be a character (\code{"covariate1"}) or a
character vector (\code{c("covariate1", "covariate2")}).
}}

\item{log_x}{Determine whether the x-axis should be displayed on a logarithmic scale}

\item{log_y}{Determine whether the y-axis should be displayed on a logarithmic scale}

\item{lifelihoodData}{\code{lifelihoodData} object created with \code{\link[=lifelihoodData]{lifelihoodData()}}.}

\item{min_sample_size}{The minimum number of individuals alive
at the beggining of a time interval for computing the observed mortality rate}

\item{rate_df}{Dataframe with mortality rate, obtained via \code{\link[=mortality_rate_data]{mortality_rate_data()}}}

\item{use_facet}{Use facet_wrap to plot one panel per group (default=FALSE)}
}
\value{
a ggplot2 plot

a ggplot2 plot

a ggplot2 plot
}
\description{
Useful function for creating a good-quality line graph
of changes in the predicted mortality rate.

If you want more control over the style of the graph,
use the \code{\link[=pred_mortality_rate]{pred_mortality_rate()}} function to retrieve data.

Useful function for creating a good-quality line graph
of changes in the empirical mortality rate.

You don't need to fit with \code{\link[=lifelihood]{lifelihood()}} to use this
function, only to retrieve a lifelihood data object
with \code{\link[=lifelihoodData]{lifelihoodData()}}

If you want more control over the style of the graph,
use the \code{\link[=compute_mortality_rate]{compute_mortality_rate()}} function to retrieve data.

Convenient function used in \code{\link[=plot_observed_mortality_rate]{plot_observed_mortality_rate()}}
and \code{\link[=plot_fitted_mortality_rate]{plot_fitted_mortality_rate()}}.
}
\details{
This function requires \href{https://ggplot2.tidyverse.org/}{ggplot2} to be installed.

This function requires \href{https://ggplot2.tidyverse.org/}{ggplot2} to be installed.
}
\keyword{internal}
