% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\name{pred_mortality_rate}
\alias{pred_mortality_rate}
\title{Compute and visualize predicted mortality rate}
\usage{
pred_mortality_rate(lifelihoodResults, newdata, intervals)
}
\arguments{
\item{lifelihoodResults}{Output object of the \code{\link[=lifelihood]{lifelihood()}} function.}

\item{newdata}{Dataframe with the values of the covariates to predict the mortality rate}

\item{intervals}{Vector of the intervals (x-axis, same unit as the data) to compute the mortality rate}
}
\description{
Compute the predicted mortality rate using the fitted values of the expected death and the survival shape.
}
