% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_mortality_rate.R
\name{pred_mortality_rate}
\alias{pred_mortality_rate}
\title{Compute predicted mortality rate}
\usage{
pred_mortality_rate(
  lifelihoodResults,
  interval_width,
  newdata = NULL,
  max_time = NULL,
  bygroup = TRUE
)
}
\arguments{
\item{interval_width}{The interval width used to calculate the mortality rate. For instance, if the time unit for deaths in the original dataset is days and \code{interval_width} is set to 10, the mortality rate will be calculated every 10 days for each group.}

\item{max_time}{The maximum time for calculating the mortality rate. If set to NULL, the time of the last observed death is used.}
}
\value{
A dataframe with 3 columns: Interval (time interval, based on \code{interval_width} value), Group (identifier of a given subgroup, or "Overall" if bygroup = FALSE), and MortalityRate (mortality rate at this time).
}
\description{
Calculate the predicted mortality rate over a given interval.
}
