% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.LifelihoodResults}
\title{Prediction with lifelihood}
\usage{
\method{predict}{LifelihoodResults}(
  lifelihoodResults,
  metric,
  newdata = NULL,
  type = c("link", "response"),
  se.fit = FALSE
)
}
\arguments{
\item{newdata}{jeu de données avec une ou plusieurs lignes à prédire, avec les mêmes nom de variables. Si facteur, il faut les mêmes les niveaux de facteurs (renvoyer erreur si inconnu: pas présent de base dans les données d'entrainement/initiales).}

\item{type}{Si "link": format de lifelihood, si "response": format original.}
}
\value{
prediction
}
\description{
Prediction with lifelihood
}
