% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.LifelihoodResults}
\title{prediction}
\usage{
\method{predict}{LifelihoodResults}(object, newdata = NULL, type = c("link", "response"), se.fit = FALSE)
}
\description{
prediction
}
