% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_event_interval_dt}
\alias{prob_event_interval_dt}
\title{Probability of dying between t and t+dt afer living until t.}
\usage{
prob_event_interval_dt(t, dt, param1, param2, family)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{dt}{Interval}

\item{param1}{Numeric. The expected longevity or time to maturity.}

\item{param2}{Numeric. The second parameter returned by lifelihood.}

\item{family}{One of "exp", "wei", "gam", "lgn"}
}
\value{
The survival probability (numeric)
}
\description{
Probability of dying between t and t+dt afer living until t.
}
\keyword{internal}
