% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_file.R
\name{read_formula}
\alias{read_formula}
\title{Read formula from config file}
\usage{
read_formula(config, parameter)
}
\arguments{
\item{parameter}{name of the parameter}
}
\value{
Formula
}
\description{
Read formula from config file
}
\keyword{internal}
