% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_output_from_file}
\alias{read_output_from_file}
\title{Read and format the output file of the program}
\usage{
read_output_from_file(file_path, covariates = NULL, path_config, MCMC)
}
\arguments{
\item{file_path}{Location of the output file of the program.}

\item{covariates}{Vector containing the names of the covariates.}

\item{path_config}{A character string specifying the file path
to the YAML configuration file.}

\item{MCMC}{Perform MCMC sampling of the parameter after
convergence to estimate their 95\% confidence interval.}
}
\value{
An object of class \code{lifelihoodResults} with all results
from the output file.
}
\description{
Takes the file path of the output file and read the
results using parsers from \code{\link[=parse_output]{parse_output()}}.
}
\keyword{internal}
