% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_output_from_file}
\alias{read_output_from_file}
\title{Functions used to read the output file and create an object with results}
\usage{
read_output_from_file(file_path, group_by_group = FALSE)
}
\arguments{
\item{file_path}{location of the output file from the programme}

\item{group_by_group}{boolean indicating whether the analysis should be made group by group or not (default to false)}
}
\description{
takes the file path of the output file and read the results using parsers from \code{\link[=parse_output]{parse_output()}}
}
