% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_output_from_file}
\alias{read_output_from_file}
\title{Read and format the output file of the program}
\usage{
read_output_from_file(file_path, group_by_group = FALSE, covariates = NULL)
}
\arguments{
\item{file_path}{Location of the output file of the program}

\item{group_by_group}{Boolean option to fit the full factorail model with all the interactions between each of the factors}
}
\value{
An object of class \code{lifelihoodResults} with all results from the output file
}
\description{
Takes the file path of the output file and read the
results using parsers from \code{\link[=parse_output]{parse_output()}}.
}
\keyword{internal}
