% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_output_from_file}
\alias{read_output_from_file}
\title{Read the output of the Lifelihood program}
\usage{
read_output_from_file(file_path)
}
\arguments{
\item{file_path}{Path to the output file}
}
\value{
A list with the following elements:
\itemize{
\item \code{datafile} Path to the output file
\item \code{seeds} A vector of four integers representing the seeds used in the simulation
\item \code{likelihood_max} The maximum likelihood value
\item \code{effects} A data frame with the following columns:
\itemize{
\item \code{Name} The name of the effect
\item \code{Estimate} The estimated value of the effect
\item \code{StdError} The standard error of the estimate
}
\item \code{parameter_ranges} A data frame with the following columns:
\itemize{
\item \code{Name} The name of the parameter
\item \code{Min} The minimum value of the parameter
\item \code{Max} The maximum value of the parameter
}
}
\description{
Read the output of the Lifelihood program
}
\examples{
# path to input
file_path = file.path(
  'data',
  'raw_data',
  'DataPierrick_GroupbyGroup',
  '100\%mort_Pierrick211genoparinteraction.out'
)

# read the output and print the results
fitted = read_output_from_file(file_path)
fitted$datafile
fitted$seeds
fitted$likelihood_max
head(fitted$effects)
head(fitted$parameter_ranges)
}
