% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lifelihood_console.R
\name{run_lifelihood_console}
\alias{run_lifelihood_console}
\title{Run lifelihood program}
\usage{
run_lifelihood_console(
  infile = "100\%mort_Pierrick211genoparinteraction.txt",
  customfile = "custom.txt",
  GbyG = 0,
  MCMC = 0,
  interval = 25,
  SEcal = 0,
  saveprobevent = 0,
  fitness = 0,
  r = 0,
  seed1 = 12,
  seed2 = 13,
  seed3 = 14,
  seed4 = 15,
  ntr = 2,
  nst = 2,
  To = 50,
  Tf = 1,
  climbrate = 1,
  precision = 0.001
)
}
\arguments{
\item{customfile}{seed used so that random values can found again (https://r-coder.com/set-seed-r/)}

\item{n_strains}{Number of strains (e.g. number of strains used for the experiment)}

\item{n_replicates}{Number of replicates per type of cross}

\item{ref_hatch_rate}{Reference hatch rate for the intra-strain cross with strain 1 (ie. female strain 1 x male strain 1)}

\item{maternal_genvar_females}{Variance among female effects (due to both maternal and genetic effects)}

\item{genvar_males}{Variance among male effects (due to genetic effects)}

\item{heterosis}{Heterosis effect in embryos from inter-strain crosses}

\item{ci_level}{Average level of cytoplasmic incompatibility}

\item{var_ci}{Variance in cytoplasmic incompatibility among Wolbachia variants}

\item{var_within}{residual error}
}
\value{
dataset with the simulated hatch rate
}
\description{
Run lifelihood program in console mode
}
\examples{
run_lifelihood_console(infile="100\%mort_Pierrickintercept.txt",ntr=10,To=50,Tf=50,seed2=34)
}
