% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_event}
\alias{simulate_event}
\title{Simulation for a single life event}
\usage{
simulate_event(object, ev, newdata)
}
\arguments{
\item{object}{Output of \code{\link[=lifelihood]{lifelihood()}}}

\item{ev}{A character of the event (must be one of "mortality",
"reproduction" or "maturity")}

\item{newdata}{An optional dataset used for prediction}
}
\description{
Internal function used to simulate one of the
life history event (maturity, reproduction, mortality).
}
\keyword{internal}
