% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_life_history}
\alias{simulate_life_history}
\title{Simulate outcomes from a fitted lifelihood model}
\usage{
simulate_life_history(
  object,
  event = c("all", "mortality", "reproduction", "maturity"),
  newdata = NULL,
  visits = NULL,
  seed = NULL,
  remove_exact_clutch_dates = TRUE
)
}
\arguments{
\item{object}{A fitted \code{lifelihoodResults} object created either with \code{\link[=lifelihood]{lifelihood()}} or
\code{\link[=create_simulation_input]{create_simulation_input()}}.}

\item{event}{Character string specifying the event(s) to simulate.
Must be one of \code{"mortality"}, \code{"maturity"}, or \code{"all"} (event=\code{"reproduction"} is equivalent to \code{"all"} as maturity and mortality are needed to simulate reproduction events).
Default is \code{"all"}, which simulates all fitted events.}

\item{newdata}{Optional \code{data.frame} providing covariate values for prediction.
If \code{NULL}, the original model data are used.}

\item{visits}{Optional data frame with 2 columns: one column with the same
name as the \code{block} argument passed to \code{\link[=as_lifelihoodData]{as_lifelihoodData()}} and
one column named exactly \code{visit}. For each block, \code{visit} corresponds to the
ages where the events of individuals have been recorded. Supplying \code{visits}
enables censoring; use \code{\link[=get_visits]{get_visits()}} to derive visit data from the fitted
data, or pass a custom visit data frame.}

\item{seed}{Optional integer. If provided, sets the random seed for reproducibility.}

\item{remove_exact_clutch_dates}{Logical. Whether to remove exact simulated
clutch-date columns from the output. If \code{FALSE}, these columns are kept
and renamed to \verb{exact_clutch_date_\{i\}}. Default is \code{TRUE}.}
}
\value{
A list of \code{data.frame} with one column per simulated event.
Each column contains simulated values for that event.
}
\description{
This function generates simulated data from a fitted lifelihood model,
for one or several life history events. By default, all fitted events are simulated.
}
