% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_life_history}
\alias{simulate_life_history}
\title{Simulate outcomes from a fitted lifelihood model}
\usage{
simulate_life_history(
  object,
  event = c("all", "mortality", "reproduction", "maturity"),
  newdata = NULL,
  use_censoring = FALSE,
  visits = NULL,
  seed = NULL
)
}
\arguments{
\item{object}{A fitted \code{lifelihoodResults} object.}

\item{event}{Character string specifying the event(s) to simulate.
Must be one of \code{"mortality"}, \code{"maturity"}, or \code{"all"} (event=\code{"reproduction"} is equivalent to \code{"all"} as maturity and mortality are needed to simulate reproduction events).
Default is \code{"all"}, which simulates all fitted events.}

\item{newdata}{Optional \code{data.frame} providing covariate values for prediction.
If \code{NULL}, the original model data are used.}

\item{use_censoring}{Whether to retrieve censoring time intervals for scalar
events (\code{maturity}, \code{mortality}). For example, returns \code{mortality_start} and
\code{mortality_end} instead of only \code{mortality}. If \code{newdata} is provided and
censoring is enabled, \code{newdata} must include the block column.
In this case, it's advised to use the \code{visits} argument and provide your own
visit data. Otherwise, visits are inferred from observed event ages.}

\item{visits}{Optionnal dataframe with 2 columns: "block" (must be the same name
as passed in \code{\link[=as_lifelihoodData]{as_lifelihoodData()}} \code{block} argument) and exactly "visit".
For each block, "visit" corresponds to the ages where the events of individuals
have been recorded.}

\item{seed}{Optional integer. If provided, sets the random seed for reproducibility.}
}
\value{
A list of \code{data.frame} with one column per simulated event.
Each column contains simulated values for that event.
}
\description{
This function generates simulated data from a fitted lifelihood model,
for one or several life history events. By default, all fitted events are simulated.
}
