% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_life_history}
\alias{simulate_life_history}
\title{Simulate outcomes from a fitted lifelihood model}
\usage{
simulate_life_history(
  object,
  event = c("all", "mortality", "reproduction", "maturity"),
  newdata = NULL,
  seed = NULL
)
}
\arguments{
\item{object}{A fitted \code{lifelihoodResults} object.}

\item{event}{Character string specifying the event(s) to simulate.
Must be one of \code{"mortality"}, \code{"maturity"}, or \code{"all"} (event=\code{"reproduction"} is equivalent to \code{"all"} as maturity and mortality are needed to simulate reproduction events).
Default is \code{"all"}, which simulates all fitted events.}

\item{newdata}{Optional \code{data.frame} providing covariate values for prediction.
If \code{NULL}, the original model data are used.}

\item{seed}{Optional integer. If provided, sets the random seed for reproducibility.}
}
\value{
A list of \code{data.frame} with one column per simulated event.
Each column contains simulated values for that event.
}
\description{
This function generates simulated data from a fitted lifelihood model,
for one or several life history events. By default, all fitted events are simulated.
}
