% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{summary.lifelihoodResults}
\alias{summary.lifelihoodResults}
\title{Summary function to be used with the output of \code{\link[=lifelihood]{lifelihood()}}}
\usage{
\method{summary}{lifelihoodResults}(object, ...)
}
\arguments{
\item{object}{output of \code{\link[=lifelihood]{lifelihood()}}}

\item{...}{Ignored}
}
\description{
S3 method to display main results of the lifelihood program.
}
