% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{summary}
\alias{summary}
\alias{summary.LifelihoodResults}
\title{Custom summary function for lifelihood}
\usage{
\method{summary}{LifelihoodResults}(object, ...)
}
\arguments{
\item{object}{\code{LifelihoodResults} object from \code{\link[=read_output_from_file]{read_output_from_file()}}}
}
\description{
Creates a custom summary method for the LifelihoodResults object
}
