% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{summary}
\alias{summary}
\alias{summary.LifelihoodResults}
\title{Custom summary function}
\usage{
\method{summary}{LifelihoodResults}(object, ...)
}
\arguments{
\item{object}{\code{LifelihoodResults} object from \code{\link[=lifelihood]{lifelihood()}}}
}
\description{
Display main results of the lifelihood program:
\itemize{
\item seeds
\item likelihood
\item effects (estimation)
\item parameter ranges
}
}
