% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelihood.R
\name{summary}
\alias{summary}
\alias{summary.lifelihoodResults}
\title{Summary function to be used with the output of \code{\link[=lifelihood]{lifelihood()}}}
\usage{
\method{summary}{lifelihoodResults}(object, ...)
}
\arguments{
\item{object}{\code{lifelihoodResults} object from \code{\link[=lifelihood]{lifelihood()}}}
}
\description{
S3 method to display main results of the lifelihood program.
}
\examples{
df <- read.csv(here::here("data/fake_sample.csv"))
df$type <- as.factor(df$type)
df$geno <- as.factor(df$geno)
head(df)

clutchs <- c(
  "clutch_start1", "clutch_end1", "clutch_size1",
  "clutch_start2", "clutch_end2", "clutch_size2"
)

dataLFH <- lifelihoodData(
  df = df,
  sex = "sex",
  sex_start = "sex_start",
  sex_end = "sex_end",
  maturity_start = "mat_start",
  maturity_end = "mat_end",
  clutchs = clutchs,
  death_start = "death_start",
  death_end = "death_end",
  covariates = c("geno", "type"),
  model_specs = c("gam", "lgn", "wei")
)

results <- lifelihood(
  lifelihoodData = dataLFH,
  path_config = here::here("config.yaml"),
  seeds = c(1, 2, 3, 4),
  raise_estimation_warning = FALSE
)
summary(results)
}
