% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SurvExp}
\alias{SurvExp}
\title{Exponential survival function}
\usage{
SurvExp(t, expt_time_to_event)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{expt_time_to_event}{Numeric. The expected longevity or time to maturity.}
}
\value{
The survival probability (numeric)
}
\description{
Exponential survival function
}
\keyword{internal}
