% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SurvGam}
\alias{SurvGam}
\title{Gamma survival function}
\usage{
SurvGam(t, ExpLong, scale)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{ExpLong}{Numeric. The expected longevity}

\item{Shape}{Numeric. The shape parameter}
}
\value{
The survival probability (numeric)
}
\description{
Gamma survival function
}
\keyword{internal}
