% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SurvLgn}
\alias{SurvLgn}
\title{Log-normal survival function}
\usage{
SurvLgn(t, ExpLong, scale)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{ExpLong}{Numeric. The expected longevity}

\item{Shape}{Numeric. The shape parameter}
}
\value{
The survival probability (numeric)
}
\description{
Log-normal survival function
}
\keyword{internal}
