% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SurvWei}
\alias{SurvWei}
\title{Weibull survival function}
\usage{
SurvWei(t, expt_time_to_event, shape)
}
\arguments{
\item{t}{Numeric. The time to event}

\item{expt_time_to_event}{Numeric. The expected longevity or time to maturity.}

\item{shape}{Numeric. The shape parameter}
}
\value{
The survival probability (numeric)
}
\description{
Weibull survival function
}
\keyword{internal}
