% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\name{SurvWei}
\alias{SurvWei}
\title{Compute and visualize predicted mortality rate}
\usage{
SurvWei(t, ExpLong, Shape)
}
\description{
Compute and visualize predicted mortality rate
}
\keyword{internal}
